Feishu: normalize reaction chat type at runtime
